// src/app/Model/standardrelief.ts
export class standardrelief {
    companyId: string;
    divisionId: string;
    departmentId: string;
    fiscalYear: any;
    reliefRate: string;
    personalAllowance: any;
    disabilityAllowance: any;
    childAllowance: any;
    dependentAllowance: any;
    lockedBy: string;
    lockTs: Date;
    minimumTaxAmount: any;
    taxGrossPercent: string;
    allStaffRelief: boolean;
    branchCode: string;
  data: string;
  status: string;

}   